// src/features/events/model/app-proto/nfs.schema.ts
import { z } from 'zod';

import { baseEventSchema } from '@/features/events/model/event';

export const nfsSchema = z.object({
  file_tx: z.boolean(),
  filename: z.string(),
  id: z.number(),
  procedure: z.string(),
  status: z.string(),
  type: z.string(),
  version: z.number(),
});

export const nfsEventSchema = baseEventSchema.extend({
  event_type: z.literal('nfs'),
  app_proto: z.literal('nfs'),
  nfs: nfsSchema,
});

export type NfsEvent = z.infer<typeof nfsEventSchema>;